Time Slider: Move Windows-specific Tickbar logic to partial class file
